feat: fix review changes
