+++ clirr-ignored-differences.xml
@@ -49,7 +49,7 @@
     </difference>
     <difference>
         <className>com/google/cloud/resourcemanager/spi/v1beta1/ResourceManagerRpc</className>
-        <method>com.google.cloud.resourcemanager.spi.v1beta1.ResourceManagerRpc$ListResult listLiens(java.lang.String, java.util.Map)</method>
+        <method>com.google.cloud.resourcemanager.ListResult listLiens(java.lang.String, java.util.Map)</method>
         <differenceType>7012</differenceType>
     </difference>
 </differences>
